pyproject.toml
[tool.poetry]
name = "testingdemo"
version = "0.1.0"
description = ""
authors = ["vinyas <vinyas.malagaudanavar@briq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.3"
alembic = "^1.10.2"
pydantic = "^1.10.6"
flask-sqlalchemy = "^3.0.3"
psycopg2-binary = "^2.9.5"



[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"
coverage = "^7.2.2"
pytest-cov = "^4.0.0"
requests-mock = "^1.10.0"
pytest-mock = "^3.10.0"



[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--cov --disable-warnings --ignore=tests/*.py"
markers = [
    "creation_test",
]

[tool.coverage.run]

source = ['app']